refactor: rename new to notifications
