+++ .env.example
@@ -149,5 +149,6 @@ OAUTH_CLIENT="YOUR_CLIENT"
 # translation.
 # APP_SEARCH_TRANSLATE_LOCALES=es,de
 
-# Name of a new feature to highlight in a notification, from src/features/new.js
-# APP_NEW_FEATURE=organisationsPage
+# Name of a new feature to highlight in a notification, from those declared in
+# src/features/notifications.js
+# APP_NEW_FEATURE_NOTIFICATION=interestingNewFeature
