+++ nuxt.config.js
@@ -32,7 +32,7 @@ export default {
       // TODO: rename env vars to prefix w/ APP_, except feature toggles
       baseUrl: process.env.PORTAL_BASE_URL,
       internalLinkDomain: process.env.INTERNAL_LINK_DOMAIN,
-      newFeature: process.env.APP_NEW_FEATURE,
+      newFeatureNotification: process.env.APP_NEW_FEATURE_NOTIFICATION,
       schemaOrgDatasetId: process.env.SCHEMA_ORG_DATASET_ID,
       siteName: APP_SITE_NAME,
       search: {
